Rename fetchPosts to fetch-posts
