+++ app/posts/page.tsx
@@ -1,7 +1,7 @@
 import Image from 'next/image'
 import Link from 'next/link'
 
-import fetchPosts from './fetchPosts'
+import fetchPosts from './fetch-posts'
 import { ReadMore } from './read-more'
 
 import { Time } from '@/components/custom/time'
